refactor(tile editor): Aplicado MVC al tile editor, separando en carpetas (model, view, controller) los diferentes archivos.
