+++ maps/combined_map.overlay.json
@@ -860,16 +860,16 @@
     "wal",
     "wal",
     "",
-    "fl0",
-    "fl1",
+    "fl3",
+    "fl3",
     "fl1",
     "fl1",
     "fl1",
     "fl4",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl2",
     "fl4",
     "fl4",
@@ -982,20 +982,20 @@
     "wal",
     "wal",
     "wal",
-    "fl0",
+    "fl3",
     "",
     "fl1",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
+    "fl3",
+    "fl3",
     "fl5",
     "fl5",
-    "fl5",
-    "fl5",
+    "fl0",
+    "fl0",
     "fl0",
     "fl5",
     "fl5",
@@ -1104,19 +1104,19 @@
     "wal",
     "wal",
     "wal",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl5",
+    "fl3",
+    "fl3",
     "",
-    "",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1226,18 +1226,18 @@
     "wal",
     "wal",
     "wal",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
+    "fl3",
+    "fl3",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl5",
-    "fl5",
     "fl0",
     "fl0",
     "fl0",
@@ -1348,17 +1348,17 @@
     "wal",
     "wal",
     "wal",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl1",
+    "fl3",
+    "fl3",
     "fl1",
     "fl0",
     "fl0",
@@ -1470,18 +1470,18 @@
     "wal",
     "wal",
     "wal",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl1",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
     "fl0",
     "fl0",
@@ -1592,132 +1592,10 @@
     "wal",
     "wal",
     "wal",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl5",
-    "fl5",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "fl2",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    "",
-    ""
-  ],
-  [
-    "",
-    "",
-    "",
-    "wal",
-    "",
-    "wal",
-    "fl0",
-    "fl0",
-    "fl0",
-    "fl0",
+    "fl3",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
     "fl0",
     "fl0",
@@ -1833,12 +1711,134 @@
     "",
     "",
     "",
+    "wal",
     "",
-    "",
-    "",
+    "wal",
+    "fl3",
+    "fl3",
     "fl0",
     "fl0",
     "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl0",
+    "fl5",
+    "fl5",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "fl2",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    ""
+  ],
+  [
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
     "fl0",
     "fl0",
@@ -1958,9 +1958,9 @@
     "",
     "",
     "",
-    "fl0",
-    "fl0",
-    "fl0",
+    "fl3",
+    "fl3",
+    "fl3",
     "fl0",
     "fl0",
     "fl0",
