+++ roguelike_project/systems/editor/tiles/controller/tile_editor_controller.py
@@ -13,7 +13,7 @@ OUTLINE_SEL    = (0, 255, 0)     # seleccionado (verde)
 OUTLINE_HOVER  = (0, 220, 255)   # hover (cian)
 OUTLINE_CHOICE = (255, 255, 0)   # elección actual (amarillo)
 
-class TileEditor:
+class TileEditorController:
     """
     • Contorno verde  → tile seleccionado
     • Contorno cian   → tile bajo el cursor
@@ -21,7 +21,7 @@ class TileEditor:
     """
     def __init__(self, state, editor_state):
         self.state   = state
-        self.editor  = editor_state     # instancia de TileEditorState
+        self.editor  = editor_state     # instancia de TileEditorControllerState
         self.picker = TilePicker(state, editor_state)
         self.toolbar = TileToolbar(state, editor_state)
 
@@ -84,64 +84,6 @@ class TileEditor:
         # cambiar al brush para empezar a pintar
         self.editor.current_tool = "brush"
 
-    def render(self, screen):
-        if not self.editor.active:
-            return
-
-        # 1) Toolbar y toggle view
-        self.toolbar.render(screen)
-
-        # 2) Picker de tiles
-        self.picker.render(screen)
-
-        # 3) Panel de ViewTile si está activo
-        if self.editor.view_active:
-            self.render_view_panel(screen)
-
-        # 4) Contornos en el mundo (hover y seleccion)
-        cam = self.state.camera
-        hover = self._tile_under_mouse(pygame.mouse.get_pos())
-        if hover:
-            rect = pygame.Rect(cam.apply((hover.x, hover.y)), cam.scale((TILE_SIZE, TILE_SIZE)))
-            pygame.draw.rect(screen, OUTLINE_HOVER, rect, 3)
-        sel = self.editor.selected_tile
-        if sel:
-            rect = pygame.Rect(cam.apply((sel.x, sel.y)), cam.scale((TILE_SIZE, TILE_SIZE)))
-            pygame.draw.rect(screen, OUTLINE_SEL, rect, 3)    
-        
-
-    def render_view_panel(self, screen):
-        panel_w = TILE_SIZE + 40
-        panel_h = 3 * (TILE_SIZE + 30)
-        x0 = self.toolbar.x + self.toolbar.size + 20
-        y0 = self.toolbar.y
-
-        panel = pygame.Surface((panel_w, panel_h), pygame.SRCALPHA)
-        panel.fill((20, 20, 20, 200))
-        font = pygame.font.SysFont("Arial", 14)
-        mouse_pos = pygame.mouse.get_pos()
-
-        items = [
-            ("Hovered",  self._tile_under_mouse(mouse_pos), OUTLINE_HOVER),
-            ("Selected", self.editor.selected_tile,           OUTLINE_SEL),
-            ("Choice",   None,                                 OUTLINE_CHOICE),
-        ]
-        for idx, (label, tile, color) in enumerate(items):
-            ty = idx * (TILE_SIZE + 30) + 10
-            if label == "Choice" and self.editor.current_choice:
-                sprite = load_image(self.editor.current_choice, (TILE_SIZE, TILE_SIZE))
-            else:
-                sprite = tile.sprite if tile else None
-
-            if sprite:
-                panel.blit(sprite, ((panel_w - TILE_SIZE)//2, ty))
-            rect = pygame.Rect((panel_w - TILE_SIZE)//2, ty, TILE_SIZE, TILE_SIZE)
-            pygame.draw.rect(panel, color, rect, 3)
-
-            text = font.render(label, True, (255, 255, 255))
-            panel.blit(text, (5, ty + TILE_SIZE + 2))
-
-        screen.blit(panel, (x0, y0))
 
     def _tile_under_mouse(self, mouse_pos):
         mx, my = mouse_pos
