+++ roguelike_project/systems/editor/tiles/controller/tools/tile_toolbar.py
@@ -5,7 +5,7 @@ from roguelike_project.utils.loader import load_image
 
 class TileToolbar:
     """
-    Barra de herramientas para el TileEditor:
+    Barra de herramientas para el TileEditorController:
       - select
       - brush
       - eyedropper
@@ -38,20 +38,7 @@ class TileToolbar:
         # Rects para detectar clicks
         self.icon_rects: dict[str, pygame.Rect] = {}
 
-    def render(self, screen):
-        """Dibuja la toolbar en pantalla."""
-        for idx, tool in enumerate(self.TOOLS):
-            px = self.x
-            py = self.y + idx * (self.size + self.padding)
-            rect = pygame.Rect(px, py, self.size, self.size)
-            self.icon_rects[tool] = rect
-
-            # icono
-            screen.blit(self.icons[tool], (px, py))
-
-            # borde: resaltado si está activa
-            color = (255, 200, 0) if self.editor.current_tool == tool else (255, 255, 255)
-            pygame.draw.rect(screen, color, rect, 4)
+ 
 
     def handle_click(self, mouse_pos) -> bool:
         for tool, rect in self.icon_rects.items():
